Start adding crystaline page
It seems like we have a lot of duplication of classes across app pages.
E.g. page-header--molevent is identical to the one for peaksel and
crystaline. So why not give them generic names that aren't app-specific?
Started to do this with Crystaline - the rest of the pages stay unchanged
for now.